Fixing usage for getInteger which should return an int
git-svn-id: https://www.silverpeas.org/svn/silverpeas/components/trunk@3021 a8e77078-a1c7-4fa5-b8fc-53c5178a176c

+++ kmelia/kmelia-war/src/main/java/com/silverpeas/kmelia/RssLastPublicationsServlet.java
@@ -117,8 +117,8 @@ public class RssLastPublicationsServlet extends HttpServlet {
 
   public Collection<PublicationDetail> getElements(KmeliaTransversal kmeliaTransversal,
       UserFull user, String spaceId) throws RemoteException {
-    int maxAge = Long.valueOf(settings.getInteger("max.age.last.publication", 0L)).intValue();
-    int nbReturned = Long.valueOf(settings.getInteger("max.nb.last.publication", 10L)).intValue();
+    int maxAge = settings.getInteger("max.age.last.publication", 0);
+    int nbReturned = settings.getInteger("max.nb.last.publication", 10);
     return kmeliaTransversal.getUpdatedPublications(spaceId, maxAge, nbReturned);
   }
